# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Star Wars Character Database API",
    "description": "API documentation for managing Star Wars characters and filtering them by collection.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://star-wars-characters-api.onrender.com"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API home route",
        "responses": {
          "200": {
            "description": "API welcome message"
          }
        }
      }
    },
    "/characters": {
      "get": {
        "summary": "Get all characters",
        "description": "Returns a list of all Star Wars characters.",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "post": {
        "summary": "Create a new character",
        "description": "Adds a new Star Wars character.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Character created successfully"
          }
        }
      }
    },
    "/characters/{id}": {
      "get": {
        "summary": "Get a character by ID",
        "description": "Returns a single Star Wars character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Character ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Character not found"
          }
        }
      },
      "put": {
        "summary": "Update a character by ID",
        "description": "Updates an existing Star Wars character.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Character ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Character updated successfully"
          },
          "404": {
            "description": "Character not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a character by ID",
        "description": "Deletes a Star Wars character.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Character ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character deleted successfully"
          },
          "404": {
            "description": "Character not found"
          }
        }
      }
    },
    "/characters/collection/{collection}": {
      "get": {
        "summary": "Get characters by collection",
        "description": "Returns characters filtered by collection such as Jedi, Sith, Droids, or Bounty Hunters.",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "description": "Collection name",
            "schema": {
              "type": "string",
              "example": "Jedi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CharacterInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Luke Skywalker"
          },
          "species": {
            "type": "string",
            "example": "Human"
          },
          "homeworld": {
            "type": "string",
            "example": "Tatooine"
          },
          "affiliation": {
            "type": "string",
            "example": "Jedi Order"
          },
          "collection": {
            "type": "string",
            "example": "Jedi"
          }
        },
        "required": ["name", "species", "affiliation", "collection"]
      }
    }
  }
}